feat: extract templates and add new ones
